The provided commit message does not match the changes given in the diff. Referencing the correct change, the commit message should be:
Commit message:
Update README with environment variables and usage details

The README has been extended to include a section on setting up environment variables for the project, as well as a usage section explaining how to train the model and generate text using the provided notebooks. It also provides a link to download model weights that are too large to be saved in the repo.